reorganize location of source files
split internal and fileio specific files in two directories
